+++ tests/Agent/IntegrationTests/IntegrationTests/RemoteServiceFixtures/BasicMvcApplicationTestFixture.cs
@@ -304,14 +304,14 @@ namespace NewRelic.Agent.IntegrationTests.RemoteServiceFixtures
         public void GetCustomAttributes(string key1, string value1, string key2, string value2)
         {
             var address = $"http://{DestinationServerName}:{Port}/Default/CustomParameters?key1={key1}&value1={value1}&key2={key2}&value2={value2}";
-            ResponseBody = DownloadStringAndAssertContains(address, "<html>");
+            ResponseBody = DownloadStringAndAssertContains(address, "Worked");
 
         }
 
         public void StartAgent()
         {
             var address = $"http://{DestinationServerName}:{Port}/Default/StartAgent";
-            ResponseBody = DownloadStringAndAssertContains(address, "<html>");
+            ResponseBody = DownloadStringAndAssertContains(address, "Worked");
         }
 
         private string GetXNewRelicId()
@@ -368,7 +368,7 @@ namespace NewRelic.Agent.IntegrationTests.RemoteServiceFixtures
         public void SimulateLostTransaction()
         {
             var address = $"http://{DestinationServerName}:{Port}/Default/SimulateLostTransaction";
-            DownloadStringAndAssertContains(address, "<html>");
+            DownloadStringAndAssertContains(address, "Worked");
         }
 
         public void GetRedis()
@@ -427,8 +427,16 @@ namespace NewRelic.Agent.IntegrationTests.RemoteServiceFixtures
 
         public void GetBackgroundThreadWithError()
         {
-            var address = $"http://{DestinationServerName}:{Port}/CustomInstrumentationAsync/GetBackgroundThreadWithError";
-            DownloadStringAndAssertEqual(address, "Worked");
+            try
+            {
+                var address = $"http://{DestinationServerName}:{Port}/CustomInstrumentationAsync/GetBackgroundThreadWithError";
+                DownloadStringAndAssertEqual(address, "Worked");
+            }
+            catch (WebException)
+            {
+                // This is expected behavior.  We need to catch and swallow this exception here to
+                // keep it from bubbling up to the test framework and failing the test.
+            }
         }
 
         public string GetBrowserTimingHeader()
@@ -443,7 +451,7 @@ namespace NewRelic.Agent.IntegrationTests.RemoteServiceFixtures
         public string GetHtmlWithCallToGetBrowserTimingHeader()
         {
             var address = $"http://{DestinationServerName}:{Port}/Default/GetHtmlWithCallToGetBrowserTimingHeader";
-            ResponseBody = DownloadStringAndAssertContains(address, "NREUM");
+            ResponseBody = DownloadStringAndAssertContains(address, "Worked");
 
             return ResponseBody;
         }